1. load all instruments 2. Load 15 mins candles by instrument (2 years old data). 3. Load candle updates by instruments.

+++ src/main/java/com/dtech/kitecon/repository/InstrumentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InstrumentRepository extends JpaRepository<Instrument, Integer> {
+public interface InstrumentRepository extends JpaRepository<Instrument, Long> {
 }
